Blind Acoustic Source Separation through binary mask approach
A Matlab code to solve a blind acoustic source separation problem with a binary mask approach in the time-frequency domain; the method is based on the assumption of sparsity of the source's energy contribution in this domain and does not take into account reverberation.
In the end, to put it in simple terms, it takes as input the sum of three signals (voices and songs) and output the three sources separated in three different audio files, with a decent quality.